requirements.txt
streamlit>=1.32
langgraph>=0.2
langchain>=0.3
langchain-community>=0.3
pydantic>=2.0
pydantic-settings>=2.0
python-dotenv>=1.0
httpx>=0.27
pillow>=10.0
selenium>=4.0
webdriver-manager>=4.0
pyautogui>=0.9
pyperclip>=1.8
duckduckgo-search>=6.0
instagrapi>=2.0
